V.0.1 test

+++ pyRPlink/src/pyRPlink.py
@@ -38,56 +38,63 @@ def main():
     print("    Ultimo dipendente: {} {}".format(dipendenti.getLastNomeDipendente(),dipendenti.getCodici()[-1].rfidCode))
     
     fileTransactionsPath = 'TRANSACTIONS.TXT'
+    importazioneIncompleta = False
     try:
         filesecs = os.path.getmtime(fileTransactionsPath)
         fileLenght = os.path.getsize(fileTransactionsPath)
         if filesecs and fileLenght:
             dataFile = strftime('%d.%m.%Y %H:%M', localtime(filesecs))
             print("Le timbrature del {} non sono ancora state importate tutte.".format(dataFile))
-            raise stop_error
+            importazioneIncompleta = True
+        if filesecs and not(fileLenght):
+            os.remove(fileTransactionsPath)
     except FileNotFoundError:
-        os.remove(fileTransactionsPath)
         pass
     
-    print("Controllo timbrature...")
-    for ip in ['192.168.29.212','192.168.29.213','192.168.39.212']:
-        transactions = TIMBRATURE(ip, fileTransactionsPath)
+    if not(importazioneIncompleta):
+        #scarico nuove timbrature
+        print("Controllo timbrature...")
+        for ip in ['192.168.29.212','192.168.29.213','192.168.39.212']:
+            transactions = TIMBRATURE(ip, fileTransactionsPath)
     
     registrazioni = []
     errori = []
     
     logPath = "log"
     
-    try:
-        ftr = open(fileTransactionsPath, "r")
+    if importazioneIncompleta:
+        #riprovo la conversione dei codici
+        try:
+            ftr = open(fileTransactionsPath, "r")
+            
+            for line in ftr:
+                timbratura = REGISTRAZIONE(line)
+                dipendente = dipendenti.findDipendenteByRfid(timbratura.getCode())
+                if dipendente:
+                    registrazioni.append(timbratura.getConvertedString(dipendente.getAScode()))
+                else:
+                    errori.append(line)
+                    
+            ftr.close()
         
-        for line in ftr:
-            timbratura = REGISTRAZIONE(line)
-            dipendente = dipendenti.findDipendenteByRfid(timbratura.getCode())
-            if dipendente:
-                registrazioni.append(timbratura.getConvertedString(dipendente.getAScode()))
-            else:
-                errori.append(line)
-                
-        ftr.close()
-    
-    
-        print("Creo il file da importare...", end=' ')
-        timbPath = "TIMBRATURE.TXT"
-        timbrature = open(timbPath, "w")
-        print("OK")
-    
-        if len(errori):
-            print("Il file dipendenti non era completo. Completalo e riprova.", end=' ')
-        else:
-            print("Converto i codici dipendente e scrivo il file da importare...", end=' ')
-            timbrature.write("\n".join(registrazioni))
-            timbrature.close()
-            shutil.move(fileTransactionsPath, os.path.join(logPath,"{}.txt".format(str(datetime.datetime.utcnow()).replace(':',''))))
+        
+            print("Creo il file da importare...", end=' ')
+            timbPath = "TIMBRATURE.TXT"
+            timbrature = open(timbPath, "w")
             print("OK")
-            print("File TIMBRATURE.TXT pronto da importare.")
-    except OSError:
-        print("Nessuna nuova timbratura scaricata.")
+        
+            if len(errori):
+                print("Il file dipendenti non era completo. Completalo e riprova.", end=' ')
+            else:
+                print("Converto i codici dipendente e scrivo il file da importare...", end=' ')
+                timbrature.write("\n".join(registrazioni))
+                timbrature.write("\n")
+                timbrature.close()
+                shutil.move(fileTransactionsPath, os.path.join(logPath,"{}.txt".format(str(datetime.datetime.utcnow()).replace(':',''))))
+                print("OK")
+                print("File TIMBRATURE.TXT pronto da importare.")
+        except OSError:
+            print("Nessuna nuova timbratura scaricata.")
         
     raise stop_error
